Rename GaitGeneratorGait classes to ModifiableGait classes.
